@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,43 +26,43 @@ dependencies:
 
   # Aesthetics
   animations: ^2.0.3
-  google_fonts: ^3.0.1
-  awesome_dialog: ^2.2.1
+  google_fonts: ^5.1.0
+  awesome_dialog: ^3.1.0
   page_transition: ^2.0.5
   cupertino_icons: ^1.0.5
   fluentui_system_icons: ^1.1.170
 
   # Widgets
   pinput: ^2.2.10
-  flutter_svg: ^1.1.0
-  rate_my_app: ^1.1.3
+  flutter_svg: ^2.0.7
+  rate_my_app: ^2.0.0
   intl_phone_field: ^3.1.0
 
   # Features
-  intl: ^0.17.0
+  intl: ^0.18.1
   clipboard: ^0.1.3
-  share_plus: ^4.0.7
+  share_plus: ^7.0.2
   url_launcher: ^6.1.2
-  image_cropper: ^2.0.3
-  image_picker: ^0.8.5+3
-  flutter_slidable: ^2.0.0
+  image_cropper: ^5.0.0
+  image_picker: ^1.0.1
+  flutter_slidable: ^3.0.0
   image_downloader: ^0.31.0
   permission_handler: ^10.0.0
-  visibility_detector: ^0.3.3
+  visibility_detector: ^0.4.0+2
   flutter_markdown: ^0.6.10+2
   cached_network_image: ^3.2.1
   random_string_generator: ^2.0.0
   full_screen_image_null_safe: ^2.0.0
   flutter_phone_direct_caller: ^2.1.1
-  internet_connection_checker: ^0.0.1+4
+  internet_connection_checker: ^1.0.0+1
 
   # Database Utilities
   provider: ^6.0.3
-  firebase_auth: ^3.3.19
-  firebase_core: ^1.17.1
-  cloud_firestore: ^3.1.17
+  firebase_auth: ^4.7.1
+  firebase_core: ^2.15.0
+  cloud_firestore: ^4.8.3
   onesignal_flutter: ^3.3.1
-  firebase_storage: ^10.2.17
+  firebase_storage: ^11.2.5
 
 dev_dependencies:
   flutter_test:
@@ -72,7 +72,7 @@ dev_dependencies:
   flutter_lints: ^2.0.1
 
   # Launcher Icon
-  flutter_launcher_icons: ^0.9.3
+  flutter_launcher_icons: ^0.13.1
 
   # Native Launch Screen
   flutter_native_splash: ^2.2.3
