@@ -1,37 +1,41 @@
 // Flutter Imports
 import 'package:flutter/material.dart';
 // Dependency Imports
+import 'package:intl_phone_field/countries.dart';
+import 'package:intl_phone_field/phone_number.dart';
 import 'package:intl_phone_field/intl_phone_field.dart';
 // File Imports
 
 class CustomFields extends StatelessWidget {
-  final TextInputType keyboardType;
-  final int maxLines;
-  final Function validator;
-  final Function onChanged; // Common
+  final TextInputType? keyboardType;
+  final int? maxLines;
+  final FormFieldValidator<String>? validator;
+  final Function(String)? onChanged;
+  final Function(PhoneNumber)? onChangedPhone;
   final dynamic initialValue;
-  final String hintText;
-  final bool obscureText;
-  final bool readOnly;
-  final String type;
+  final String? hintText;
+  final bool? obscureText;
+  final bool? readOnly;
+  final String? type;
   // Phone Number Field Variables
-  final bool autoFocus;
-  final FocusNode focusNode;
-  final String phoneIsoCode;
-  final String nonInternationalNumber;
-  final Function onCountryChanged;
+  final bool? autoFocus;
+  final FocusNode? focusNode;
+  final String? phoneIsoCode;
+  final String? nonInternationalNumber;
+  final Function(Country)? onCountryChanged;
 
   const CustomFields({
-    Key key,
+    Key? key,
     this.keyboardType,
     this.maxLines,
     this.validator,
-    this.onChanged, // Common
+    this.onChanged,
+    this.onChangedPhone,
     this.initialValue,
     this.hintText,
     this.obscureText,
     this.readOnly,
-    @required this.type,
+    required this.type,
     // Phone Number Field Variables
     this.autoFocus,
     this.focusNode,
@@ -116,7 +120,7 @@ class CustomFields extends StatelessWidget {
                     ),
                   ),
                 ),
-                onChanged: onChanged,
+                onChanged: onChangedPhone,
                 onCountryChanged: onCountryChanged,
                 initialCountryCode: phoneIsoCode,
                 initialValue: nonInternationalNumber,
