@@ -2,10 +2,10 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 // Chat Room Model
 class ChatRoom {
-  final String chatRoomId;
-  final Timestamp connectionDate;
-  final String customerUID;
-  final String professionalUID;
+  final String? chatRoomId;
+  final Timestamp? connectionDate;
+  final String? customerUID;
+  final String? professionalUID;
 
   ChatRoom({
     this.chatRoomId,
