@@ -1,11 +1,11 @@
 // Service Model
 class Help4YouServices {
-  final String serviceId;
-  final String serviceTitle;
-  final String serviceDescription;
-  final double servicePrice;
-  final bool visibility;
-  final String professionalId;
+  final String? serviceId;
+  final String? serviceTitle;
+  final String? serviceDescription;
+  final double? servicePrice;
+  final bool? visibility;
+  final String? professionalId;
 
   Help4YouServices({
     this.serviceId,
