@@ -1,6 +1,6 @@
 // Help4You Unique User ID
 class Help4YouUser {
-  final String uid;
+  final String? uid;
 
   Help4YouUser({
     this.uid,
@@ -9,15 +9,15 @@ class Help4YouUser {
 
 // Help4You Professional Details
 class UserDataProfessional {
-  final String uid;
-  final String fullName;
-  final String occupation;
-  final String profilePicture;
-  final String countryCode;
-  final String phoneIsoCode;
-  final String nonInternationalNumber;
-  final String phoneNumber;
-  final String status;
+  final String? uid;
+  final String? fullName;
+  final String? occupation;
+  final String? profilePicture;
+  final String? countryCode;
+  final String? phoneIsoCode;
+  final String? nonInternationalNumber;
+  final String? phoneNumber;
+  final String? status;
 
   UserDataProfessional({
     this.uid,
