@@ -11,7 +11,7 @@ import 'package:h4y_partner/constants/signature_button.dart';
 import 'package:h4y_partner/constants/custom_text_field.dart';
 
 class DeleteAccPhoneAuthScreen extends StatefulWidget {
-  const DeleteAccPhoneAuthScreen({Key key}) : super(key: key);
+  const DeleteAccPhoneAuthScreen({Key? key}) : super(key: key);
 
   @override
   DeleteAccPhoneAuthScreenState createState() =>
@@ -22,7 +22,7 @@ class DeleteAccPhoneAuthScreenState extends State<DeleteAccPhoneAuthScreen> {
   // Text Field Variables
   String countryCode = "+1";
   String phoneIsoCode = "US";
-  String nonInternationalNumber;
+  String? nonInternationalNumber;
 
   @override
   Widget build(BuildContext context) {
@@ -66,7 +66,7 @@ class DeleteAccPhoneAuthScreenState extends State<DeleteAccPhoneAuthScreen> {
                   autoFocus: true,
                   phoneIsoCode: phoneIsoCode,
                   nonInternationalNumber: nonInternationalNumber,
-                  onChanged: (phone) {
+                  onChangedPhone: (phone) {
                     setState(() {
                       nonInternationalNumber = phone.number;
                     });
