@@ -6,15 +6,15 @@ import 'package:google_fonts/google_fonts.dart';
 // File Imports
 
 class Pages extends StatelessWidget {
-  final String graphicImage;
-  final String title;
-  final String description;
+  final String? graphicImage;
+  final String? title;
+  final String? description;
 
   const Pages({
-    Key key,
-    @required this.graphicImage,
-    @required this.title,
-    @required this.description,
+    Key? key,
+    required this.graphicImage,
+    required this.title,
+    required this.description,
   }) : super(key: key);
 
   @override
@@ -27,11 +27,11 @@ class Pages extends StatelessWidget {
         children: [
           SizedBox(
             height: MediaQuery.of(context).size.height * 0.35,
-            child: SvgPicture.asset(graphicImage),
+            child: SvgPicture.asset(graphicImage!),
           ),
           const SizedBox(height: 20.0),
           Text(
-            title,
+            title!,
             style: GoogleFonts.balooPaaji2(
               height: 1.5,
               fontSize: 26.0,
@@ -42,7 +42,7 @@ class Pages extends StatelessWidget {
           ),
           const SizedBox(height: 10.0),
           Text(
-            description,
+            description!,
             style: GoogleFonts.balooPaaji2(
               height: 1.3,
               fontSize: 20.0,
