@@ -12,33 +12,33 @@ import 'package:h4y_partner/services/database.dart';
 import 'package:h4y_partner/constants/custom_snackbar.dart';
 
 class IconButtonStream extends StatelessWidget {
-  final Help4YouUser user;
-  final File imageFile;
-  final GlobalKey<FormState> formKey;
-  final String countryCode;
-  final String nonInternationalNumber;
-  final String fullName;
-  final String occupation;
-  final String phoneIsoCode;
+  final Help4YouUser? user;
+  final File? imageFile;
+  final GlobalKey<FormState>? formKey;
+  final String? countryCode;
+  final String? nonInternationalNumber;
+  final String? fullName;
+  final String? occupation;
+  final String? phoneIsoCode;
 
   const IconButtonStream({
-    Key key,
-    @required this.user,
-    @required this.imageFile,
-    @required this.formKey,
-    @required this.countryCode,
-    @required this.nonInternationalNumber,
-    @required this.fullName,
-    @required this.occupation,
-    @required this.phoneIsoCode,
+    Key? key,
+    required this.user,
+    required this.imageFile,
+    required this.formKey,
+    required this.countryCode,
+    required this.nonInternationalNumber,
+    required this.fullName,
+    required this.occupation,
+    required this.phoneIsoCode,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return StreamBuilder(
-      stream: DatabaseService(uid: user.uid).userData,
+      stream: DatabaseService(uid: user!.uid).userData,
       builder: (context, snapshot) {
-        UserDataProfessional userData = snapshot.data;
+        UserDataProfessional? userData = snapshot.data as UserDataProfessional?;
         return IconButton(
           icon: const Icon(
             CupertinoIcons.checkmark_alt,
@@ -51,28 +51,28 @@ class IconButtonStream extends StatelessWidget {
               if (imageFile != null) {
                 Reference firebaseStorageRef = FirebaseStorage.instance
                     .ref()
-                    .child(("H4Y Profile Pictures/${user.uid}"));
-                UploadTask uploadTask = firebaseStorageRef.putFile(imageFile);
+                    .child(("H4Y Profile Pictures/${user!.uid}"));
+                UploadTask uploadTask = firebaseStorageRef.putFile(imageFile!);
                 await uploadTask;
                 String downloadAddress =
                     await firebaseStorageRef.getDownloadURL();
-                await DatabaseService(uid: user.uid)
+                await DatabaseService(uid: user!.uid)
                     .updateProfilePicture(downloadAddress);
               } else {
-                await DatabaseService(uid: user.uid)
-                    .updateProfilePicture(userData.profilePicture);
+                await DatabaseService(uid: user!.uid)
+                    .updateProfilePicture(userData!.profilePicture);
               }
             }
 
-            if (countryCode.contains("+")) {
-              countryCode.replaceAll("+", "");
+            if (countryCode!.contains("+")) {
+              countryCode!.replaceAll("+", "");
             }
             HapticFeedback.heavyImpact();
             FocusScope.of(context).unfocus();
             try {
-              if (formKey.currentState.validate()) {
+              if (formKey!.currentState!.validate()) {
                 String phoneNumber = "+$countryCode$nonInternationalNumber";
-                if (userData.phoneNumber != phoneNumber) {
+                if (userData!.phoneNumber != phoneNumber) {
                   await AuthService().phoneAuthentication(
                     fullName,
                     occupation,
@@ -83,24 +83,23 @@ class IconButtonStream extends StatelessWidget {
                     "Update Phone Number",
                     context,
                   );
-                  await DatabaseService(uid: user.uid).updateUserData(
-                    fullName ?? userData.fullName,
-                    occupation ?? userData.occupation,
-                    phoneNumber ?? userData.phoneNumber,
-                    countryCode ?? userData.countryCode,
-                    phoneIsoCode ?? userData.phoneIsoCode,
-                    nonInternationalNumber ?? userData.nonInternationalNumber,
+                  await DatabaseService(uid: user!.uid).updateUserData(
+                    fullName,
+                    occupation,
+                    phoneNumber,
+                    countryCode,
+                    phoneIsoCode,
+                    nonInternationalNumber,
                   );
                 } else {
-                  await DatabaseService(uid: user.uid)
+                  await DatabaseService(uid: user!.uid)
                       .updateUserData(
-                        fullName ?? userData.fullName,
-                        occupation ?? userData.occupation,
-                        userData.phoneNumber ?? userData.phoneNumber,
-                        userData.countryCode ?? userData.countryCode,
-                        userData.phoneIsoCode ?? userData.phoneIsoCode,
-                        userData.nonInternationalNumber ??
-                            userData.nonInternationalNumber,
+                        fullName,
+                        occupation,
+                        userData.phoneNumber,
+                        userData.countryCode,
+                        userData.phoneIsoCode,
+                        userData.nonInternationalNumber,
                       )
                       .then(
                         (value) => Navigator.pop(context),
