@@ -22,7 +22,7 @@ import 'package:h4y_partner/screens/delete_account_screens/delete_phone_auth_scr
 import 'package:h4y_partner/screens/personal_data_screen/components/icon_button_stream.dart';
 
 class PersonalDataScreen extends StatefulWidget {
-  const PersonalDataScreen({Key key}) : super(key: key);
+  const PersonalDataScreen({Key? key}) : super(key: key);
 
   @override
   PersonalDataScreenState createState() => PersonalDataScreenState();
@@ -30,33 +30,33 @@ class PersonalDataScreen extends StatefulWidget {
 
 class PersonalDataScreenState extends State<PersonalDataScreen> {
   // Text Field Variables
-  String fullName;
-  String occupation;
-  String countryCode;
-  String nonInternationalNumber;
-  String phoneIsoCode;
-  String profilePicture;
+  String? fullName;
+  String? occupation;
+  String? countryCode;
+  String? nonInternationalNumber;
+  String? phoneIsoCode;
+  String? profilePicture;
 
   // Global Key
   final formKey = GlobalKey<FormState>();
 
   // Active Image File
-  File imageFile;
+  File? imageFile;
 
   // Select Image Via Image Picker
   Future getImage(ImageSource source) async {
     final selected = await ImagePicker().pickImage(source: source);
     if (selected == null) return;
-    File image = File(selected.path);
+    File? image = File(selected.path);
     image = await cropImage(selected);
     setState(() {
-      imageFile = image;
+      imageFile = image!;
     });
   }
 
   // Crop Selected Image
-  Future<File> cropImage(XFile selectedFile) async {
-    CroppedFile cropped = await ImageCropper().cropImage(
+  Future<File?> cropImage(XFile selectedFile) async {
+    CroppedFile? cropped = await ImageCropper().cropImage(
       sourcePath: selectedFile.path,
       aspectRatio: const CropAspectRatio(
         ratioX: 1.0,
@@ -64,6 +64,7 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
       ),
       cropStyle: CropStyle.rectangle,
     );
+    // ignore: unnecessary_null_comparison
     if (cropped == null) return null;
     return File(cropped.path);
   }
@@ -73,13 +74,15 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
     super.initState();
     FirebaseFirestore.instance
         .collection("H4Y Users Database")
-        .doc(FirebaseAuth.instance.currentUser.uid)
+        .doc(FirebaseAuth.instance.currentUser!.uid)
         .get()
         .then((value) {
       setState(() {
-        countryCode = value.data()["Country Code"];
-        phoneIsoCode = value.data()["Phone ISO Code"];
-        nonInternationalNumber = value.data()["Non International Number"];
+        fullName = value.data()!["Full Name"];
+        occupation = value.data()!["Occupation"];
+        countryCode = value.data()!["Country Code"];
+        phoneIsoCode = value.data()!["Phone ISO Code"];
+        nonInternationalNumber = value.data()!["Non International Number"];
       });
     });
   }
@@ -87,10 +90,10 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
   @override
   Widget build(BuildContext context) {
     // Get User
-    final user = Provider.of<Help4YouUser>(context);
+    final user = Provider.of<Help4YouUser?>(context);
 
-    if (countryCode.contains("+")) {
-      countryCode = countryCode.replaceAll("+", "");
+    if (countryCode!.contains("+")) {
+      countryCode = countryCode!.replaceAll("+", "");
     }
 
     return Scaffold(
@@ -133,9 +136,10 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
                   bottom: 10.0,
                 ),
                 child: StreamBuilder(
-                  stream: DatabaseService(uid: user.uid).userData,
+                  stream: DatabaseService(uid: user!.uid).userData,
                   builder: (context, snapshot) {
-                    UserDataProfessional userData = snapshot.data;
+                    UserDataProfessional? userData =
+                        snapshot.data as UserDataProfessional?;
                     if (snapshot.hasData) {
                       return Column(
                         children: [
@@ -153,12 +157,12 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
                                         ? Container(
                                             decoration: BoxDecoration(
                                               image: DecorationImage(
-                                                image: FileImage(imageFile),
+                                                image: FileImage(imageFile!),
                                               ),
                                             ),
                                           )
                                         : CachedNetworkImage(
-                                            imageUrl: userData.profilePicture,
+                                            imageUrl: userData!.profilePicture!,
                                             fit: BoxFit.fill,
                                           ),
                                   ),
@@ -172,7 +176,7 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
                                     child: GestureDetector(
                                       onTap: () {
                                         Widget dialogButton(String title,
-                                            Color color, Function onTap) {
+                                            Color color, VoidCallback onTap) {
                                           return Padding(
                                             padding: const EdgeInsets.symmetric(
                                               horizontal: 15.0,
@@ -209,7 +213,7 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
                                         AwesomeDialog(
                                           context: context,
                                           headerAnimationLoop: false,
-                                          dialogType: DialogType.INFO,
+                                          dialogType: DialogType.info,
                                           body: Column(
                                             children: [
                                               dialogButton(
@@ -311,9 +315,9 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
                               type: "Normal",
                               keyboardType: TextInputType.name,
                               hintText: "Enter Full Name...",
-                              initialValue: userData.fullName,
-                              validator: (String value) {
-                                if (value.isEmpty) {
+                              initialValue: fullName,
+                              validator: (String? value) {
+                                if (value!.isEmpty) {
                                   return "Name field cannot be empty";
                                 } else if (value.length < 2) {
                                   return "Name must be atleast 2 characters long";
@@ -338,7 +342,7 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
                               stream: FirebaseFirestore.instance
                                   .collection("H4Y Occupation Database")
                                   .snapshots(),
-                              builder: (context, snapshot) {
+                              builder: (context, AsyncSnapshot snapshot) {
                                 if (snapshot.hasData) {
                                   List<DropdownMenuItem> occupationItems = [];
                                   for (int i = 0;
@@ -352,14 +356,15 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
                                         child: Text(
                                           snap['Occupation'],
                                           style: const TextStyle(
-                                              color: Colors.black),
+                                            color: Colors.black,
+                                          ),
                                         ),
                                       ),
                                     );
                                   }
                                   return CustomDropdown(
                                     hintText: 'Select Occupation...',
-                                    value: userData.occupation,
+                                    value: occupation,
                                     validator: (occupationValue) {
                                       if (occupationValue.isEmpty) {
                                         return "Please select an occupation";
@@ -419,10 +424,10 @@ class PersonalDataScreenState extends State<PersonalDataScreen> {
                                 const EdgeInsets.symmetric(horizontal: 15.0),
                             child: CustomFields(
                               type: "Phone",
-                              phoneIsoCode: userData.phoneIsoCode,
+                              phoneIsoCode: userData!.phoneIsoCode,
                               nonInternationalNumber:
                                   userData.nonInternationalNumber,
-                              onChanged: (phone) {
+                              onChangedPhone: (phone) {
                                 setState(() {
                                   nonInternationalNumber = phone.number;
                                 });
