@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 // File Imports
 
 class DashedLine extends StatelessWidget {
-  const DashedLine({Key key}) : super(key: key);
+  const DashedLine({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
