@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 // File Imports
 
 class Widget2Body extends StatelessWidget {
-  const Widget2Body({Key key}) : super(key: key);
+  const Widget2Body({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
