@@ -9,17 +9,17 @@ import 'package:h4y_partner/services/auth.dart';
 import 'package:h4y_partner/constants/signature_button.dart';
 
 class UpdateNumVerificationScreen extends StatefulWidget {
-  final String phoneIsoCode;
-  final String nonInternationalNumber;
-  final String phoneNumber;
-  final Function submitOTP;
+  final String? phoneIsoCode;
+  final String? nonInternationalNumber;
+  final String? phoneNumber;
+  final Function(String)? submitOTP;
 
   const UpdateNumVerificationScreen({
-    Key key,
-    @required this.phoneIsoCode,
-    @required this.nonInternationalNumber,
-    @required this.phoneNumber,
-    @required this.submitOTP,
+    Key? key,
+    required this.phoneIsoCode,
+    required this.nonInternationalNumber,
+    required this.phoneNumber,
+    required this.submitOTP,
   }) : super(key: key);
 
   @override
@@ -30,9 +30,9 @@ class UpdateNumVerificationScreen extends StatefulWidget {
 class UpdateNumVerificationScreenState
     extends State<UpdateNumVerificationScreen> {
   // Text Field Variable
-  String fullName;
-  String phoneIsoCode;
-  String nonInternationalNumber;
+  String? fullName;
+  String? phoneIsoCode;
+  String? nonInternationalNumber;
 
   // Pin Put Declarations
   Color borderColor = const Color.fromRGBO(114, 178, 238, 1);
@@ -98,7 +98,7 @@ class UpdateNumVerificationScreenState
                   focusedPinTheme: defaultPinTheme.copyWith(
                     width: 63,
                     height: 68,
-                    decoration: defaultPinTheme.decoration.copyWith(
+                    decoration: defaultPinTheme.decoration!.copyWith(
                       border: Border.all(color: borderColor),
                     ),
                   ),
@@ -108,7 +108,7 @@ class UpdateNumVerificationScreenState
                 GestureDetector(
                   onTap: () async {
                     FirebaseAuth.instance.verifyPhoneNumber(
-                      phoneNumber: widget.phoneNumber,
+                      phoneNumber: widget.phoneNumber!,
                       timeout: const Duration(seconds: 120),
                       verificationCompleted:
                           (PhoneAuthCredential credential) async {},
@@ -120,7 +120,7 @@ class UpdateNumVerificationScreenState
                         );
                       },
                       codeSent:
-                          (String verificationId, int resendToken) async {},
+                          (String? verificationId, int? resendToken) async {},
                       codeAutoRetrievalTimeout: (String verificationId) async {
                         verificationId = verificationId;
                       },
